The Dummy App
This app demonstrates the heart reading capability that could not be tested due to database issues. The app reads a textile that is created by a .sh script that also reconnects the Bluetooth Module HC-05 after dropping connection each time. The .sh file is run in the terminal, creates and writes data from the Bluetooth Module to the data.txt.